refactor: rename files to kebab-case and update theme config components
