+++ package.json
@@ -135,7 +135,7 @@
     "cmdk": "0.2.0",
     "culori": "^4.0.2",
     "flagged": "3.0.0",
-    "framer-motion": "^12.0.3",
+    "framer-motion": "^12.6.3",
     "fuse.js": "7.0.0",
     "himalaya": "^1.1.0",
     "i18next": "^24.2.1",
@@ -168,8 +168,8 @@
     "@testing-library/react": "^16.2.0",
     "@types/lodash": "^4.17.14",
     "@types/node": "^20.17.11",
-    "@types/react": "^18.3.1",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
     "@typescript-eslint/eslint-plugin": "^8.21.0",
     "@typescript-eslint/parser": "^8.21.0",
     "@vitejs/plugin-react-swc": "^3.7.2",
@@ -192,7 +192,7 @@
     "react-dom": "19.1.0",
     "react-router-dom": "^7.1.3",
     "shadcn": "2.3.0",
-    "tailwindcss": "^3.4.11",
+    "tailwindcss": "^3.4.17",
     "tailwindcss-animate": "^1.0.7",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
