+++ src/core/components/sidepanels/panels/theme-configuration/border-radius-input.tsx
@@ -1,5 +1,6 @@
 import { Slider } from "@/ui/shadcn/components/ui/slider";
 import { useThrottledCallback } from "@react-hookz/web";
+import { useState } from "react";
 
 type BorderRadiusInputProps = {
   value: string;
@@ -8,16 +9,19 @@ type BorderRadiusInputProps = {
 };
 
 const BorderRadiusInput = ({ value, onChange, disabled }: BorderRadiusInputProps) => {
-  const throttledChange = useThrottledCallback((value: string) => onChange(value), [value], 200, true);
-
+  const [_value, _setValue] = useState(value);
+  const throttledChange = useThrottledCallback(onChange, [value], 200, true);
   return (
     <Slider
       min={0}
       step={1}
       max={50}
       disabled={disabled}
-      value={[Number(value.replace("px", ""))]} 
-      onValueChange={(value) => throttledChange(value[0].toString())}
+      value={[Number(_value.replace("px", ""))]}
+      onValueChange={(value) => {
+        _setValue(value[0].toString() + "px");
+        throttledChange(value[0].toString());
+      }}
       className="flex-1 cursor-pointer"
     />
   );
