+++ src/core/components/sidepanels/panels/theme-configuration/color-picker-input.tsx
