+++ src/core/components/sidepanels/panels/theme-configuration/index.ts
@@ -1,5 +1,5 @@
-import BorderRadiusInput from "@/core/components/sidepanels/panels/theme-configuration/BorderRadiusInput";
-import ColorPickerInput from "@/core/components/sidepanels/panels/theme-configuration/ColorPickerInput";
+import BorderRadiusInput from "@/core/components/sidepanels/panels/theme-configuration/border-radius-input";
+import ColorPickerInput from "@/core/components/sidepanels/panels/theme-configuration/color-picker-input";
 import FontSelector from "@/core/components/sidepanels/panels/theme-configuration/font-selector";
 
 export { BorderRadiusInput, ColorPickerInput, FontSelector };
